refactor: controlpanel to use tailwind
